Speed improvements
- introduced readQualityBuffer for speed improvements
- previously 3x of lofreq2, now 2x
- see experiments/2021-01-14-speedtest-readqualitybuffer.txt
- bumped Nim version to 1.4.2
- Changed all BAM args to -b (from -i)
- MQ255 now translates to high(int)

+++ src/lofreqpkg/utils.nim
@@ -23,11 +23,11 @@ proc refSymbolAtIndel*(reverse: bool): char =# evil hack to support strand for "
     REF_SYMBOL_AT_INDEL_FW
 
 
-proc decodeQual*(encqual: char, offset: uint = 33): uint =
-  return uint(ord(encqual))-offset
+proc decodeQual*(encqual: char, offset: uint8 = 33): uint8 =
+  return uint8(ord(encqual))-offset
 
 
-proc encodeQual*(qual: uint, offset: uint = 33): char =
+proc encodeQual*(qual: uint8, offset: uint8 = 33): char =
   assert qual+offset < 255
   result = chr(qual+offset)
